flytt kode som kun trengs i EksternKlageController ut i egen EksternKlageService og dropp wrapper-funksjon for oppretting av klage
