+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/klage/KlageService.kt
@@ -10,34 +10,15 @@ import no.nav.familie.kontrakter.felles.klage.Stønadstype
 import no.nav.familie.ks.sak.common.exception.Feil
 import no.nav.familie.ks.sak.common.exception.FunksjonellFeil
 import no.nav.familie.ks.sak.integrasjon.familieintegrasjon.IntegrasjonClient
-import no.nav.familie.ks.sak.integrasjon.tilbakekreving.TilbakekrevingKlient
-import no.nav.familie.ks.sak.kjerne.behandling.BehandlingService
-import no.nav.familie.ks.sak.kjerne.behandling.domene.Behandling
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.VedtakService
-import no.nav.familie.ks.sak.kjerne.fagsak.FagsakService
 import no.nav.familie.ks.sak.kjerne.fagsak.domene.Fagsak
-import no.nav.familie.ks.sak.kjerne.klage.dto.OpprettKlageDto
 import org.springframework.stereotype.Service
 import java.time.LocalDate
 
 @Service
 class KlageService(
-    private val fagsakService: FagsakService,
     private val klageClient: KlageClient,
     private val integrasjonClient: IntegrasjonClient,
-    private val behandlingService: BehandlingService,
-    private val vedtakService: VedtakService,
-    private val tilbakekrevingKlient: TilbakekrevingKlient,
 ) {
-    fun opprettKlage(
-        fagsakId: Long,
-        opprettKlageDto: OpprettKlageDto,
-    ) {
-        val fagsak = fagsakService.hentFagsak(fagsakId)
-
-        opprettKlage(fagsak, opprettKlageDto.kravMottattDato)
-    }
-
     fun opprettKlage(
         fagsak: Fagsak,
         kravMottattDato: LocalDate,
@@ -70,27 +51,5 @@ class KlageService(
                 ?: throw Feil("Fikk ikke fagsakId=$fagsakId tilbake fra kallet til klage.")
 
         return klagerPåFagsak.map { it.brukVedtaksdatoFraKlageinstansHvisOversendt() }
-    }
-
-    fun hentFagsystemVedtak(fagsakId: Long): List<FagsystemVedtak> {
-        val behandlinger = behandlingService.hentFerdigstilteBehandlinger(fagsakId)
-
-        val ferdigstilteKsVedtak = behandlinger.map { tilFagsystemVedtak(it) }
-        val vedtakTilbakekreving = tilbakekrevingKlient.hentTilbakekrevingsvedtak(fagsakId)
-
-        return ferdigstilteKsVedtak + vedtakTilbakekreving
-    }
-
-    private fun tilFagsystemVedtak(behandling: Behandling): FagsystemVedtak {
-        val vedtak = vedtakService.hentAktivVedtakForBehandling(behandling.id)
-
-        return FagsystemVedtak(
-            eksternBehandlingId = behandling.id.toString(),
-            behandlingstype = behandling.type.visningsnavn,
-            resultat = behandling.resultat.displayName,
-            vedtakstidspunkt = vedtak.vedtaksdato ?: error("Mangler vedtakstidspunkt for behandling=${behandling.id}"),
-            fagsystemType = FagsystemType.ORDNIÆR,
-            regelverk = behandling.kategori.tilRegelverk(),
-        )
     }
 }
